@@ -5,7 +5,7 @@
     "keywords": ["translator", "nette"],
     "license": "MIT",
     "require": {
-        "php": ">= 7.1",
+        "php": ">= 7.4",
         "nette/utils": "^3.0|^4.0",
         "nette/finder": "^2.5|^3.0",
         "nette/neon": "^3.0",
