@@ -9,17 +9,28 @@ use Efabrica\Translatte\Resolver\ChainResolver;
 use Efabrica\Translatte\Resource\NeonDirectoryResource;
 use Efabrica\Translatte\Translator;
 use Latte\Engine;
-use Nette\Application\UI\ITemplateFactory;
+use Nette\Application\UI\TemplateFactory;
 use Nette\DI\CompilerExtension;
 use Nette\DI\Definitions\FactoryDefinition;
+use Nette\DI\Definitions\ServiceDefinition;
 use Nette\DI\Definitions\Statement;
 use Nette\DI\DynamicParameter;
-use Nette\DI\ServiceDefinition;
 use Nette\PhpGenerator\PhpLiteral;
 use Nette\Schema\Expect;
 use Nette\Schema\Schema;
 
-class TranslationExtension extends CompilerExtension
+/**
+ * @phpstan-type TranslationConfig object{
+ *     default: string,
+ *     fallback: array<string>,
+ *     dirs: array<string>,
+ *     cache: Statement|DynamicParameter|null,
+ *     resolvers: array<Statement>,
+ *     resources: array<Statement>,
+ *     recordTranslate: Statement|DynamicParameter|null
+ * }
+ */
+final class TranslationExtension extends CompilerExtension
 {
     public function getConfigSchema(): Schema
     {
@@ -38,30 +49,33 @@ class TranslationExtension extends CompilerExtension
     {
         $builder = $this->getContainerBuilder();
 
+        /** @var TranslationConfig $config */
+        $config = $this->config;
+
         // Prepare params for translator
-        $params = ['defaultLang' => $this->config->default];
-        if (!empty($this->config->resolvers)) {
-            $params['resolver'] = new Statement(ChainResolver::class, [$this->config->resolvers]);
+        $params = ['defaultLang' => $config->default];
+        if (!empty($config->resolvers)) {
+            $params['resolver'] = new Statement(ChainResolver::class, [$config->resolvers]);
         }
-        if ($this->config->cache) {
-            $params['cache'] = $this->config->cache;
+        if ($config->cache !== null) {
+            $params['cache'] = $config->cache;
         }
-        if ($this->config->recordTranslate) {
-            $params['recordTranslate'] = $this->config->recordTranslate;
+        if ($config->recordTranslate !== null) {
+            $params['recordTranslate'] = $config->recordTranslate;
         }
 
         $translator = $builder->addDefinition($this->prefix('translator'))
             ->setFactory(Translator::class, $params);
 
         // Configure translator
-        foreach ($this->config->resources as $resource) {
+        foreach ($config->resources as $resource) {
             $translator->addSetup('addResource', [$resource]);
         }
-        if (!empty($this->config->fallback)) {
-            $translator->addSetup('setFallbackLanguages', [$this->config->fallback]);
+        if (!empty($config->fallback)) {
+            $translator->addSetup('setFallbackLanguages', [$config->fallback]);
         }
-        if (!empty($this->config->dirs)) {
-            $translator->addSetup('addResource', [new Statement(NeonDirectoryResource::class, [$this->config->dirs])]);
+        if (!empty($config->dirs)) {
+            $translator->addSetup('addResource', [new Statement(NeonDirectoryResource::class, [$config->dirs])]);
         }
     }
 
@@ -72,7 +86,7 @@ class TranslationExtension extends CompilerExtension
         /** @var ServiceDefinition $translator */
         $translator = $builder->getDefinition($this->prefix('translator'));
 
-        $templateFactoryName = $builder->getByType(ITemplateFactory::class);
+        $templateFactoryName = $builder->getByType(TemplateFactory::class);
         if ($templateFactoryName !== null) {
             /** @var ServiceDefinition $templateFactory */
             $templateFactory = $builder->getDefinition($templateFactoryName);
@@ -88,7 +102,7 @@ class TranslationExtension extends CompilerExtension
             $latteFactory->getResultDefinition()
                 ->addSetup('addProvider', ['translator', $builder->getDefinition($this->prefix('translator'))]);
 
-            if (version_compare(Engine::VERSION, '3', '<')) {
+            if (version_compare(Engine::VERSION, '3', '<')) { // @phpstan-ignore-line
                 $latteFactory->getResultDefinition()->addSetup('?->onCompile[] = function($engine) { ?::install($engine->getCompiler()); }', ['@self', new PhpLiteral(TranslateMacros::class)]);
             }
         }
