@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace Efabrica\Translatte\Record;
 
-class NullRecord implements RecordInterface
+final class NullRecord implements RecordInterface
 {
     public function save(string $message): void
     {
